# GODRIVER-2734 Benchmark and reduce allocations in bson.Raw.String